refactor: combine traits for code reuse
