+++ app/Filament/Helpers/PairsAirlinesWithStands.php
@@ -3,11 +3,8 @@
 namespace App\Filament\Helpers;
 
 use App\Filament\Resources\TranslatesStrings;
-use App\Models\Aircraft\Aircraft;
 use Carbon\Carbon;
 use Closure;
-use Filament\Forms\Components\Select;
-use Filament\Forms\Components\TextInput;
 use Filament\Forms\Components\TimePicker;
 use Filament\Tables\Actions\DetachAction;
 use Filament\Tables\Columns\TextColumn;
@@ -15,66 +12,23 @@ use Illuminate\Support\Facades\DB;
 
 trait PairsAirlinesWithStands
 {
+    use HasAirlinePairingCommonFields;
     use TranslatesStrings;
 
-    private static string $defaultColumnValue = '--';
-
-    private static function commonPairingTableColumns(): array
+    private static function airlineStandPairingTableColumns(): array
     {
         return [
-            TextColumn::make('aircraft_id')
-                ->default(self::$defaultColumnValue)
-                ->label(self::translateTablePath('columns.aircraft'))
-                ->formatStateUsing(fn (int|string $state) => is_int($state) ? Aircraft::find($state)->code : '')
-                ->sortable()
-                ->searchable(),
-            TextColumn::make('destination')
-                ->label(self::translateTablePath('columns.destination'))
-                ->default(self::$defaultColumnValue)
-                ->sortable(),
-            TextColumn::make('full_callsign')
-                ->default(self::$defaultColumnValue)
-                ->label(self::translateTablePath('columns.full_callsign')),
-            TextColumn::make('callsign_slug')
-                ->default(self::$defaultColumnValue)
-                ->label(self::translateTablePath('columns.callsign_slug')),
-            TextColumn::make('priority')
-                ->default(self::$defaultColumnValue)
-                ->label(self::translateTablePath('columns.priority')),
+            ...self::commonPairingTableColumns(),
             TextColumn::make('not_before')
                 ->label(self::translateTablePath('columns.not_before'))
                 ->date('H:i')
         ];
     }
 
-    private static function commonPairingFormFields(): array
+    private static function airlineStandPairingFormFields(): array
     {
         return [
-            Select::make('aircraft_id')
-                ->searchable()
-                ->options(SelectOptions::aircraftTypes())
-                ->label(self::translateFormPath('aircraft.label'))
-                ->helperText(self::translateFormPath('aircraft.helper')),
-            TextInput::make('destination')
-                ->label(self::translateFormPath('destination.label'))
-                ->helperText(self::translateFormPath('destination.helper'))
-                ->maxLength(4),
-            TextInput::make('full_callsign')
-                ->label(self::translateFormPath('full_callsign.label'))
-                ->helperText(self::translateFormPath('full_callsign.helper'))
-                ->maxLength(4),
-            TextInput::make('callsign_slug')
-                ->label(self::translateFormPath('callsign_slug.label'))
-                ->helperText(self::translateFormPath('callsign_slug.helper'))
-                ->maxLength(4),
-            TextInput::make('priority')
-                ->label(self::translateFormPath('priority.label'))
-                ->helperText(self::translateFormPath('priority.helper'))
-                ->default(100)
-                ->numeric()
-                ->minValue(1)
-                ->maxValue(9999)
-                ->required(),
+            ...self::commonPairingFormFields(),
             TimePicker::make('not_before')
                 ->label(self::translateFormPath('not_before.label'))
                 ->helperText(self::translateFormPath('not_before.helper'))
