+++ app/Filament/Resources/TerminalResource/RelationManagers/AirlinesRelationManager.php
@@ -36,7 +36,7 @@ class AirlinesRelationManager extends RelationManager
                     ->label(self::translateTablePath('columns.icao'))
                     ->sortable()
                     ->searchable(),
-                ...self::commonPairingTableColumns(),
+                ...self::airlineTerminalPairingTableColumns(),
             ])
             ->headerActions([
                 AttachAction::make('pair-airline')
@@ -44,12 +44,12 @@ class AirlinesRelationManager extends RelationManager
                         $action->getRecordSelect()
                             ->label(self::translateFormPath('icao.label'))
                             ->required(),
-                        ...self::commonPairingFormFields(),
+                        ...self::airlineTerminalPairingFormFields(),
                     ])
             ])
             ->actions([
                 EditAction::make('edit-airline-pairing')
-                    ->form(self::commonPairingFormFields()),
+                    ->form(self::airlineTerminalPairingFormFields()),
                 DetachAction::make('unpair-airline')
                     ->label(self::translateFormPath('remove.label'))
                     ->using(self::unpairingClosure())
